fix: set /live and /ready as first middleware
To fix the error of Error [ERR_HTTP_HEADERS_SENT]:
Cannot set headers after they are sent to the client
I've move liveness and readiness from the end of the middlwares stack
to the front. So with this
solution I can bypass the express way of handling routes.
Now we can enable GracefulServer from anywhere.
Before this fix, we need to setup GracefulServer before the express
endpoints and middleware.